+++ src/GameRules.ts
@@ -1,12 +1,14 @@
 import { CELL_SIZE, Vector } from "./Commons";
 
 export interface CellType {
+        isWalkable: () => boolean,
         render: (ctx: CanvasRenderingContext2D, pos: Vector) => void
 }
 
 const EPS = 1 / CELL_SIZE;
 
 const bedrock = {
+        isWalkable: () => false,
         render: (ctx: CanvasRenderingContext2D, pos: Vector) => {
                 ctx.fillStyle = '#ffc080';
                 ctx.fillRect(pos.x, pos.y, 1 + EPS, 1 + EPS);
@@ -14,6 +16,7 @@ const bedrock = {
 };
 
 const black = {
+        isWalkable: () => true,
         render: (ctx: CanvasRenderingContext2D, pos: Vector) => {
                 ctx.fillStyle = '#222222';
                 ctx.fillRect(pos.x, pos.y, 1 + EPS, 1 + EPS);
@@ -21,11 +24,15 @@ const black = {
 };
 
 const snow = {
+        isWalkable: () => true,
         render: (ctx: CanvasRenderingContext2D, pos: Vector) => {
                 ctx.fillStyle = '#ffffff';
                 ctx.fillRect(pos.x, pos.y, 1 + EPS, 1 + EPS);
         },
 };
+
+export class Unit {
+}
 
 export class World {
         constructor() {
